Modify json_to_jsonlist function to keep keys of interest
feat: modified jsonl_to_jsonlist function so that it includes only the keys of interest from the JSON dict. The keys are in the keep_cols list

Previously, we were storing most of the JSON file, which made the resulting demands on memory significantly higher. This tweak leans down the process

As a next step, I need to check why the 'user' key is not showing up in the jSON dictionaries. Then accordingly, the sub-keys of user need to be included.